fix(mccpp): T-mac-30236/30154 — widen line_num to int64 so __LINE__ past INT_MAX doesn't wrap negative
#line 2147483647 then __LINE__ gave -2147483648 (int line_num wrap); clang/gcc
give 2147483648. Widen BufferedFile.line_num/line_ref + TokenString last/save_line_num
+ tok_line_num to int64_t; the two __LINE__-value emitters snprintf %lld. DWARF/gen
reads narrow losslessly (all real lines << INT_MAX) -> o0-neutral. Re-banked
fmt/census-bank (mccpp.c snprintf 67->68, one added site). Residual: macro-body
__LINE__ past INT_MAX (token-string 32-bit encoding) + #line-value clamp, noted.
TDD cli/line_macro_near_intmax (anti-vacuous vs wrap AND saturate).

+++ src/mcc.h
@@ -458,8 +458,8 @@ typedef struct BufferedFile {
 	uint8_t *buf_end;
 	int fd;
 	struct BufferedFile *prev;
-	int line_num;
-	int line_ref;
+	int64_t line_num;
+	int64_t line_ref;
 	unsigned long cst_base;
 	int ifndef_macro;
 	int ifndef_macro_saved;
@@ -481,8 +481,8 @@ typedef struct TokenString {
 	int len;
 	int need_spc;
 	int allocated_len;
-	int last_line_num;
-	int save_line_num;
+	int64_t last_line_num;
+	int64_t save_line_num;
 	struct TokenString *prev;
 	const int *prev_ptr;
 	char alloc;
